+++ packages/codemods/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.21](https://github.com/mittwald/flow/compare/1.1.20...1.1.21) (2026-09-07)
+
+**Note:** Version bump only for package @mittwald/flow-codemods
+
+## [1.1.20](https://github.com/mittwald/flow/compare/1.1.19...1.1.20) (2026-09-04)
+
+### Bug Fixes
+
+* **Option:** infer textValue from mixed children and keep value stable ([#3076](https://github.com/mittwald/flow/issues/3076)) ([e80bf3c](https://github.com/mittwald/flow/commit/e80bf3c6e97543c3d3b2db9d75811e2958e2fd65))
+
+## [1.1.19](https://github.com/mittwald/flow/compare/1.1.18...1.1.19) (2026-09-04)
+
+**Note:** Version bump only for package @mittwald/flow-codemods
+
+## [1.1.18](https://github.com/mittwald/flow/compare/1.1.17...1.1.18) (2026-09-04)
+
+**Note:** Version bump only for package @mittwald/flow-codemods
+
+## [1.1.17](https://github.com/mittwald/flow/compare/1.1.16...1.1.17) (2026-09-04)
+
+**Note:** Version bump only for package @mittwald/flow-codemods
+
+## [1.1.16](https://github.com/mittwald/flow/compare/1.1.15...1.1.16) (2026-09-04)
+
+**Note:** Version bump only for package @mittwald/flow-codemods
+
+## [1.1.15](https://github.com/mittwald/flow/compare/1.1.14...1.1.15) (2026-09-03)
+
+**Note:** Version bump only for package @mittwald/flow-codemods
+
 ## [1.1.14](https://github.com/mittwald/flow/compare/1.1.13...1.1.14) (2026-09-03)
 
 **Note:** Version bump only for package @mittwald/flow-codemods